Checkpoint for JSON data type reads
